drop logs file and use stdout
Instead of logging to non-syslog log files the script will now just log
to STDOUT. These log files were almost entirely used for development
anyway so having this output to STDOUT will actually be easier to manage
and read. This also eliminates the issue of managing where to put the log
files and cleaning up logs of files. The log level that is output is
controlled by verbosity flags.